authentication

+++ server/controllers/api.js
@@ -10,7 +10,7 @@ module.exports = class API {
         const token = req.headers.token;
         try {
             const user = await User.findOne({token:token});
-            const posts = await Post.find({user_id:user.id});
+            const posts = await Post.find({user_id:user._id});
             res.status(200).json(posts);
         } catch (error) {
             res.status(404).json({ message: error.message })
@@ -24,10 +24,11 @@ module.exports = class API {
         afterDate[2] = (parseInt(afterDate[2]) + 1).toString()
         afterDate = afterDate.join('-')
         console.log(afterDate)
+        // console.log(token);
 
         try {
             const user = await User.findOne({token:token});
-            const posts = await Post.find({user_id:user.id, created: { $gte: date, $lte: afterDate }});
+            const posts = await Post.find({user_id:user._id, created: { $gte: date, $lte: afterDate }});
 
             res.status(200).json(posts);
         } catch (error) {
@@ -40,7 +41,7 @@ module.exports = class API {
         const token = req.headers.token;
         try {
             const user = await User.findOne({token:token});
-            const post = await Post.findOne({token:user.token, id:id});
+            const post = await Post.findOne({token:user.token, _id:id});
             res.status(200).json(post);
         } catch (error) {
             res.status(404).json({ message: error.message })
@@ -51,11 +52,11 @@ module.exports = class API {
         const token = req.headers.token;
         const post = req.body;
         const user = await User.findOne({token:token});
-
+        // console.log(token);
         // const imagename = req.file.filename;
         // post.image = imagename;
-        post.user_id = user.id;
-        console.log(post)
+        post.user_id = user._id;
+        // console.log(post)
         try {
             await Post.create(post)
             res.status(201).json({ message: "Successfully create new Journal" })
@@ -71,7 +72,7 @@ module.exports = class API {
         const user = await User.findOne({token:token});
         const newPost = req.body;
         try {
-            await Post.findOneAndUpdate({token:user.token, _id:id},newPost);
+            await Post.findOneAndUpdate({user_id:user._id, _id:id},newPost);
             // await Post.findByIdAndUpdate(id, newPost);
             res.status(200).json({ message: "Successfully update your Journal" })
         } catch (error) {
@@ -81,9 +82,11 @@ module.exports = class API {
 
     static async deleteJournals(req, res) {
         const id = mongo.ObjectID(req.params.id);
+        const token = req.headers.token;
+        const user = await User.findOne({token:token});
         // const user = await User.findOne({token:token});
         try {
-            const result = await Post.findOneAndDelete({_id : id});
+            const result = await Post.findOneAndDelete({user_id:user._id, _id : id});
             // if (result.image != '') {
             //     try {
             //         fs.unlinkSync('./uploads/'+result.image);
